+++ frontend/src/components/overview/CaseOverviewFooter.tsx
